pyproject.toml
[tool.poetry]
name = "hyperbolic-agentkit-sample"
version = "0.1.0"
description = "A sample app that uses Hyperbolic AgentKit"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
cdp-langchain = "^0.0.8"
cdp-agentkit-core = "^0.0.6"
langchain-core = "^0.3.15"
langchain-openai = "^0.2.6"
langgraph = "^0.2.45"
pydantic = "^2.9.2"
cdp-sdk = "^0.12.0"
twitter-langchain = "^0.0.6"
python-dotenv = "^1.0.0"
langchain-community = "^0.3.11"
langchain-experimental = "^0.3.3"
paramiko = "^3.4.0"
duckduckgo-search = "^6.4.1"
langchain-nomic = "^0.1.4"
beautifulsoup4 = "^4.12.3"
scikit-learn = "^1.6.0"
nomic = {extras = ["local"], version = "^3.3.4"}
tiktoken = "^0.8.0"
langchain-anthropic = "^0.3.0"
chromadb = "^0.6.2"
torch = "2.2"
sentence-transformers = "^3.3.1"
google-cloud-aiplatform = "^1.38.0"
vertexai = "^0.0.1"
numpy = "<2.0.0"
gradio = "^5.12.0"
pyaudio = "0.2.14"
elevenlabs = "1.50.3"
pydub = "0.25.1"
pymilvus = {extras = ["model"], version = "^2.4.0"}
datasets = "^3.2.0"
einops = "^0.8.0"
cartesia = "^1.3.1"
deepgram-sdk = "^3.9.0"
sounddevice = "^0.5.1"
torchaudio = "~2.2.0"
backoff = "^2.2.1"
websockets = "^12.0"
aiohttp = "^3.8.5"


[tool.poetry.group.dev.dependencies]
pytest-playwright = "^0.6.2"
chromadb = "^0.6.2"
torch = "2.2"
sentence-transformers = "^3.3.1"
google-cloud-aiplatform = "^1.38.0"
vertexai = "^0.0.1"
numpy = "<2.0.0"
gradio = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
